requirements.txt
fastapi
uvicorn
pymysql
pyJWT
# 用來列出容器需要安裝的所有Python套件，相當於這個API的依賴
# Python套件 : 由多個模組、函數、類別等組成的程式庫，可以通過import在程式裡使用
# 依賴 : 一個軟體項目需要其他軟體來正常運行，換句話說，依賴就是你的專案需要使用到的套件
